requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
pytest==5.4.2
